http://issues.apache.org/jira/browse/EXTSCRIPT-111 warning fixes missing headers added
git-svn-id: https://svn.apache.org/repos/asf/myfaces/extensions/scripting/trunk@930680 13f79535-47bb-0310-9956-ffa450edef68

+++ extscript-core-root/extscript-core/src/main/java/org/apache/myfaces/scripting/core/dependencyScan/filter/ScanIdentifierFilter.java
@@ -18,7 +18,7 @@ public class ScanIdentifierFilter implements ClassFilter {
     }
 
     public boolean isAllowed(Integer identifier, String clazz) {
-        int id = identifier.intValue();
+        int id = identifier;
         for(int engineType: _engineType) {
             boolean allowed = engineType == id;
             if(allowed) return true;
